+++ R/plot.trajectory.R
@@ -59,6 +59,7 @@ trajectory_graph <- function(x, fill) {
   # find activity names
   out <- sub(".*Activity: ", "", out)
   nodes <- data.frame(label = sub(" .*", "", out), stringsAsFactors=FALSE)
+  nodes$type <- nodes$label
   nodes$shape <- "box"
   nodes$shape[c(forks, rollbacks)] <- "diamond"
   nodes$style <- "solid"
@@ -144,5 +145,23 @@ trajectory_graph <- function(x, fill) {
     }
   } else edges <- NULL
 
-  DiagrammeR::create_graph(nodes, edges)
+  DiagrammeR::create_graph(nodes, edges) %>%
+    postprocess_clones()
+}
+
+postprocess_clones <- function(graph) {
+  clones <- dplyr::filter_(graph$nodes_df, ~type == "Clone")
+  for (i in seq_len(nrow(clones))) {
+    n <- as.numeric(strsplit(clones[i,]$tooltip, "n: ", fixed=TRUE)[[1]][2])
+    id_clone <- clones[i,]$id
+    edges <- dplyr::filter_(graph$edges_df, ~from == id_clone)
+    if (n+1 <= nrow(edges))
+      graph <- DiagrammeR::delete_edge(graph, id=edges[1,]$id)
+    edges <- dplyr::filter_(graph$edges_df, ~from == id_clone)
+    while (n < nrow(edges)) {
+      graph <- DiagrammeR::delete_edge(graph, id=edges[nrow(edges),]$id)
+      edges <- dplyr::filter_(graph$edges_df, ~from == id_clone)
+    }
+  }
+  graph
 }