+++ tests/testthat/test-plot-trajectory.R
@@ -69,6 +69,29 @@ test_that("a complex trajectory is correctly converted to graph", {
              c("Seize", "Branch", "Clone", "Seize", "Timeout", "Release", "Trap",
                "Timeout", "Timeout", "SetAttribute", "Seize", "Timeout", "Release",
                "Release", "Rollback", "Synchronize", "Rollback", "Release"),
-             c(1, 2, 2, 2, 3, 3, 3, 4, 5, 6, 7, 7, 9, 10, 11, 12, 13, 14, 15, 16, 17, 17),
-             c(2, 3, 10, 16, 4, 7, 16, 5, 6, 16, 8, 9, 16, 11, 12, 13, 14, 15, 1, 17, 2, 18))
+             c(1, 2, 2, 2, 3, 3, 4, 5, 6, 7, 7, 9, 10, 11, 12, 13, 14, 15, 16, 17, 17),
+             c(2, 3, 10, 16, 4, 7, 5, 6, 16, 8, 9, 16, 11, 12, 13, 14, 15, 1, 17, 2, 18))
+})
+
+test_that("edges are removed for ignored subtrajectories in Clone", {
+  x <- trajectory() %>%
+    clone(3,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    clone(2,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    clone(1,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    clone(0,
+          trajectory() %>% timeout(1),
+          trajectory() %>% set_attribute("asdf", 1)) %>%
+    synchronize()
+
+  test_graph(x,
+             c("Clone", "Timeout", "SetAttribute", "Clone", "Timeout", "SetAttribute",
+               "Clone", "Timeout", "SetAttribute", "Clone", "Timeout", "SetAttribute", "Synchronize"),
+             c(1, 1, 1, 2, 3, 4, 4, 5, 6, 7, 8, 9, 11, 12),
+             c(2, 3, 4, 4, 4, 5, 6, 7, 7, 8, 10, 10, 13, 13))
 })